Level3 람다변형 BinayFunc적용

+++ Sparta_PersonalAssignment/src/day1_nbCalculator/ArithmeticCalculator.java
@@ -45,8 +45,9 @@ public class ArithmeticCalculator extends Calculator {
                 secondNumber = Double.parseDouble(getInput(br));
             }
         }
-        result = operatorType.apply(firstNumber, secondNumber).doubleValue();
+        result = operatorType.apply(firstNumber, secondNumber);
         addArr(result);
         return result;
     }
+
 }
